Use default asset graph keys, update all demos to reflect this
Previously the px-tree component redefined the default keys for
asset data. This is an issue because it makes it hard for apps
to share the same items data across multiple components that use
the asset graph API (context browser, tree, breadcrumbs).

This change removes that behavior and reverts to using the default
keys defined by the asset graph behavior. Demos and tests also
updated to show examples using the default keys.